add python3 support sans tests
tweaked imports and added some names and tweaked the order of some
things. the python3 interface to memcached is binary - it will store the
bytes you give it, so you would need to encode unicode objects and
decode whatever you got back if it is supposed to be a string.

TODO: make tests work with python3 as well.