add failing test for large data load
loading a bunch of keys at once fails because the .send() call doens't
deliver all the data at once so nothing is ready when listening for the
response.

there are two choices to fix this - either remove the .settimeout(5) call
and make all your sockets blocking (and then use .sendall() instead of
send), or, everywhere you call .send(), check how many bytes were
written and if it wasn't the full message send the rest of the message.